+++ lib/map.js
@@ -104,6 +104,7 @@ module.exports = function(argv, cb) {
 
     const cluster = new Cluster(pool);
     const index = new Index(pool);
+    const orphan = new Orphan(pool, argv, output);
 
     if (argv['skip-import']) {
         pool.query(`
@@ -379,7 +380,7 @@ module.exports = function(argv, cb) {
 
                                 if (!active.length) {
                                     pg_done();
-                                    return orphanAddr();
+                                    return orphans();
                                 }
                             } else {
                                 // send kill signal so the pgpool can be shut down properly
@@ -410,84 +411,18 @@ module.exports = function(argv, cb) {
         });
     }
 
-    function orphanAddr() {
-        let label;
-        if (!argv.label)
-            label = require('./label/titlecase')();
-        else
-            label = require('./label/' + argv.label)();
+    function orphans() {
+        orphanQ = Queue();
 
-        pool.connect((err, client, pg_done) => {
-            if (err) return cb(err);
-
-            client.query(`SELECT count(*) FROM address_cluster a LEFT JOIN network_cluster n ON a.id = n.address WHERE n.address IS NULL;`, (err, res) => {
-                if (err) return cb(err);
-
-                const bar = new prog('ok - Formatting Unmached Addresses [:bar] :percent :etas', {
-                    complete: '=',
-                    incomplete: ' ',
-                    width: 20,
-                    total: parseInt(res.rows[0].count)
-                });
-
-                const cursor = client.query(new Cursor(`SELECT a._text AS address_text, n._text as network_text, ST_AsGeoJSON(a.geom) AS geom FROM address_cluster a LEFT JOIN network_cluster n ON a.id = n.address WHERE n.address IS NULL;`));
-
-                return iterate();
-
-                function iterate() {
-                    cursor.read(100, (err, rows) => {
-                        if (err) return cb(err);
-
-                        if (!rows.length) {
-                            pg_done();
-                            return finalize();
-                        }
-
-                        rows.forEach((row) => {
-                            let feat = {
-                                type: 'Feature',
-                                properties: {
-                                    'carmen:text': label(row),
-                                    'carmen:center': false,
-                                    'carmen:addressnumber': []
-                                },
-                                geometry: {
-                                    type: 'GeometryCollection',
-                                    geometries: [{
-                                        type: 'MultiPoint',
-                                        coordinates: []
-                                    }]
-                                }
-                            };
-
-                            let geom = JSON.parse(row.geom);
-
-                            geom.coordinates.forEach((coord) => {
-                                if (coord[2] % 1 != 0 && argv.unitMap) {
-                                    let unit = parseInt(String(coord[2]).split('.')[1]);
-                                    let num = String(coord[2]).split('.')[0];
-
-                                    coord[2] = `${num}${argv.unitMap[unit]}`;
-                                }
-
-                                feat.properties['carmen:addressnumber'].push(coord.pop());
-                                feat.geometry.geometries[0].coordinates.push(coord);
-                            });
-
-                            feat.properties['carmen:center'] = turf.pointOnSurface(feat.geometry.geometries[0]).geometry.coordinates;
-                            feat.properties['carmen:addressnumber'] = [ feat.properties['carmen:addressnumber'] ];
-
-                            if (argv.country) feat.properties['carmen:geocoder_stack'] = argv.country;
-
-                            output.write(JSON.stringify(post.feat(feat)) + '\n');
-                        });
-
-                        bar.tick(100);
-                        return iterate();
-                    });
-                }
-            });
+        orphanQ.defer((done) => {
+            orphan.address(done)
         });
+
+        orphanQ.defer((done) => {
+            orphan.network(done)
+        });
+
+        orphan.awaitAll(finalize);
     }
